Add system_info plugin for gathering system metrics
- Implemented a Go plugin for system information retrieval, including OS details, CPU, memory, network interfaces, and storage.
- Added Python integration for beacon and session commands to request system info.
- Created necessary Go module files for the new plugin.
- Updated server configuration paths for certificates and implant data.
- Enhanced server startup process to ensure proper extraction of embedded assets and plugins.
- Added new binary files for Windows debug and release versions of the system_info plugin.
