+++ src/Server/Modules/global_objects.py
@@ -9,10 +9,10 @@ can handle bytes and strings.
 
 from .utils.content_handler import TomlFiles
 from .utils.logging import LoggingClass as Logger
-from .utils.initial_setup import generate_config_file
 from tomlkit.exceptions import InvalidCharInStringError
 
 import os
+import sys
 
 
 beacon_list = {}
@@ -20,21 +20,37 @@ command_list = {}
 sessions_list = {}
 multiplayer_connections = {}
 
+if getattr(sys, "frozen", False):
+    base_dir = os.path.expanduser("~/.PrometheanProxy")
+    plugin_dir = os.path.join(base_dir, "Plugins")
+    config_dir = base_dir
+else:
+    # Running from source: use repository's Server/Plugins and Server/config.toml
+    server_root = os.path.dirname(os.path.dirname(__file__))  # .../src/Server
+    plugin_dir = os.path.join(server_root, "Plugins")
+    # Use local config at src/Server/config.toml when running as a script
+    config_dir = server_root
 
-config_dir = os.path.expanduser("~/.PrometheanProxy/")
+# Ensure directory exists (no-op for most executable directories if not writable)
+try:
+    os.makedirs(config_dir, exist_ok=True)
+except PermissionError:
+    # If we can't create the directory next to the exe, fall back to home dir
+    if getattr(sys, "frozen", False):
+        config_dir = os.path.expanduser("~/.PrometheanProxy/")
+        os.makedirs(config_dir, exist_ok=True)
+
 config_file_path = os.path.join(config_dir, "config.toml")
 
 try:
     with TomlFiles(config_file_path) as f:
         config = f
 except FileNotFoundError:
-    generate_config_file()
     with TomlFiles(config_file_path) as f:
         config = f
 except InvalidCharInStringError:
     os.remove(config_file_path)
     print("Invalid character found in config file. Regenerating config file.")
-    generate_config_file()
     with TomlFiles(config_file_path) as f:
         config = f
