+++ src/Server/Plugins/__init__.py
@@ -1,0 +1,1 @@
+# Makes 'Server.plugins' a package for plugin discovery.
