+++ src/Server/Plugins/netstat/__init__.py
@@ -1,0 +1,1 @@
+# Package marker for plugins.netstat to enable plugin discovery.
