+++ src/Server/Plugins/system_info/__init__.py
@@ -1,0 +1,1 @@
+# Package marker for plugins.system_info to enable plugin discovery.
