+++ src/Server/config.toml
@@ -2,12 +2,12 @@
 listenaddress = "0.0.0.0"
 webPort = 8000
 port = 2000
-TLSCertificateDir = "~/Certificates/"
+TLSCertificateDir = "~/.PrometheanProxy/Certificates/"
 TLSCertificate = "cert.pem"
 TLSkey = "key.pem"
 GUI = false 
 quiet_mode = false
-implant_directory = "~/ImplantData/"
+implant_directory = "~/.PrometheanProxy/ImplantData/"
 module_location = "~/.PrometheanProxy/modules"
 
 [multiplayer]
@@ -24,8 +24,8 @@ keylength = 111
 active = false
 listenaddress = "0.0.0.0"
 port = 11001
-TLSCertificate = "~/Certificates/cert.pem"
-TLSkey = "~/Certificates/key.pem"
+TLSCertificate = "~/PrometheanProxy/Certificates/cert.pem"
+TLSkey = "~/PrometheanProxy/Certificates/key.pem"
 debugPrint = true
 
 [beacon]
