--- src/Server/plugins/__init__.py
@@ -1,1 +1,0 @@
-# Makes 'Server.plugins' a package for plugin discovery.
