--- src/Server/plugins/netstat/__init__.py
@@ -1,1 +1,0 @@
-# Package marker for plugins.netstat to enable plugin discovery.
