--- src/Server/plugins/system_info/__init__.py
@@ -1,1 +1,0 @@
-# Package marker for plugins.system_info to enable plugin discovery.
